Switch to gluster_volume module
Using the gluster_volume module allows idempotent volume options. It
also allows us to get rid of the peer probe step since the module
automatically does it as a part of volume creation. By delegating this
action to the cluster "master", we ensure the nodes get added to the
larger cluster.

Signed-off-by: John Strunk <jstrunk@redhat.com>