Add role for updating authorized ssh keys
This new role expects all keys to be *.pub
They should be placed either in the ssh-allow or ssh-deny directories
depending on whether they should be present or absent.
The intended usage is to grant someone access by placing their key in
the allow directory. If that access should be removed, their key file
should be moved to the deny directory, where it can remain indefinitely.

Signed-off-by: John Strunk <jstrunk@redhat.com>