Fix tmpdir error in tls role
When running tls role, if some hosts already have valid keys, it was
possible that the tmpdir might not get created on the local machine
because the 'run-once' machine was one w/ a valid key. This change
moves the tmpdir creation/deletion outside the conditional task block so
that the dir always gets created/removed even if it isn't used.

Fixes: #17

Signed-off-by: John Strunk <jstrunk@redhat.com>